got basic removal and traversing of lumpnodes working
