+++ addon/components/rdfa/rdfa-editor.ts
@@ -110,7 +110,10 @@ export default class RdfaEditor extends Component<RdfaEditorArgs> {
 
   getPlugins(): EditorPlugin[] {
     const pluginNames = this.plugins;
-    const plugins = [new BasicStyles(), new LumpNodePlugin()];
+    const plugins = [
+      new BasicStyles(),
+      //  new LumpNodePlugin()
+    ];
     for (const name of pluginNames) {
       const plugin = this.owner.lookup(`plugin:${name}`) as EditorPlugin | null;
       if (plugin) {
